# docs/swagger-list-genius.json
{
      "openapi": "3.0.1",
      "info": {
            "title": "List Genius API - João Dias Dev",
            "description": "NET 8.0 Web RESTFul API",
            "contact": {
                  "name": "João Dias",
                  "url": "https://github.com/joaodiasdev"
            },
            "version": "v1"
      },
      "paths": {
...
            "/api/v{version}/Group/{sortDirection}/{pageSize}/{page}": {
                  "get": {
                        "tags": [
                              "Group"
                        ],
                        "parameters": [
                              {
                                    "name": "name",
                                    "in": "query",
                                    "schema": {
                                          "type": "string"
                                    }
                              },
                              {
                                    "name": "sortDirection",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              },
                              {
                                    "name": "pageSize",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "integer",
                                          "format": "int32"
                                    }
                              },
                              {
                                    "name": "page",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "integer",
                                          "format": "int32"
                                    }
                              },
                              {
                                    "name": "version",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "Success",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/GroupVO"
                                                      }
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/GroupVO"
                                                      }
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/GroupVO"
                                                      }
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/GroupVO"
                                                      }
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "type": "array",
                                                      "items": {
                                                            "$ref": "#/components/schemas/GroupVO"
                                                      }
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "description": "Bad Request",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "description": "Unauthorized",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
            "/api/v{version}/Group/{id}": {
                  "get": {
                        "tags": [
                              "Group"
                        ],
                        "parameters": [
                              {
                                    "name": "id",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "integer",
                                          "format": "int64"
                                    }
                              },
                              {
                                    "name": "version",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "Success",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/GroupVO"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/GroupVO"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/GroupVO"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/GroupVO"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/GroupVO"
                                                }
                                          }
                                    }
                              },
                              "400": {
                                    "description": "Bad Request",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "description": "Unauthorized",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              }
                        }
                  },
                  "patch": {
                        "tags": [
                              "Group"
                        ],
                        "parameters": [
                              {
                                    "name": "id",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "integer",
                                          "format": "int64"
                                    }
                              },
                              {
                                    "name": "version",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "200": {
                                    "description": "Success"
                              }
                        }
                  },
                  "delete": {
                        "tags": [
                              "Group"
                        ],
                        "parameters": [
                              {
                                    "name": "id",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "integer",
                                          "format": "int64"
                                    }
                              },
                              {
                                    "name": "version",
                                    "in": "path",
                                    "required": true,
                                    "schema": {
                                          "type": "string"
                                    }
                              }
                        ],
                        "responses": {
                              "204": {
                                    "description": "No Content"
                              },
                              "400": {
                                    "description": "Bad Request",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              },
                              "401": {
                                    "description": "Unauthorized",
                                    "content": {
                                          "text/plain": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/json": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "application/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          },
                                          "text/xml": {
                                                "schema": {
                                                      "$ref": "#/components/schemas/ProblemDetails"
                                                }
                                          }
                                    }
                              }
                        }
                  }
            },
...
      },
      "components": {
            "schemas": {
...
                  "GroupVO": {
                        "type": "object",
                        "properties": {
                              "id": {
                                    "type": "integer",
                                    "format": "int64"
                              },
                              "name": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "image": {
                                    "type": "string",
                                    "format": "byte",
                                    "nullable": true
                              },
                              "description": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "enabled": {
                                    "type": "boolean"
                              },
                              "created_date": {
                                    "type": "string",
                                    "format": "date-time"
                              },
                              "updated_date": {
                                    "type": "string",
                                    "format": "date-time"
                              },
                              "links": {
                                    "type": "array",
                                    "items": {
                                          "$ref": "#/components/schemas/HyperMediaLink"
                                    },
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "HyperMediaLink": {
                        "type": "object",
                        "properties": {
                              "action": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "href": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "rel": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "type": {
                                    "type": "string",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": false
                  },
                  "ProblemDetails": {
                        "type": "object",
                        "properties": {
                              "type": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "title": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "status": {
                                    "type": "integer",
                                    "format": "int32",
                                    "nullable": true
                              },
                              "detail": {
                                    "type": "string",
                                    "nullable": true
                              },
                              "instance": {
                                    "type": "string",
                                    "nullable": true
                              }
                        },
                        "additionalProperties": { }
                  },
...
            }
      }
}